Prevent SSX attacks

+++ app/Http/Controllers/QuestionController.php
@@ -75,9 +75,11 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        /*
         if(Gate::denies('update-question', $question)) {
             abort(403, "Access denied");
-        }
+        } */
+        $this->authorize("update", $question);
         return view('question.edit', compact('question'));
 
     }
@@ -91,9 +93,10 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
-        if(Gate::denies('update-question', $question)) {
+        /*if(Gate::denies('update-question', $question)) {
         abort(403, "Access denied");
-        }
+        }*/
+        $this->authorize("update", $question);
         $question->update($request->only('title', 'body'));
         return redirect()->route('question.index')->with('success', "Your question has been updated");
     }
@@ -106,9 +109,11 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        /*
         if(Gate::denies('delete-question', $question)) {
             abort(403, "Access denied");
-        }
+        } */
+        $this->authorize("delete", $question);
         $question->delete();
         return redirect()->route('question.index')->with('success', "Your question has been deleted");
     }
